TestMapBuilder typo fix

+++ engine/src/main/battlecode/world/TestMapBuilder.java
@@ -45,7 +45,7 @@ public class TestMapBuilder {
     }
     
     public TestMapBuilder setPassability() {
-        this.passabilityArray = new int[width*height];
+        this.passabilityArray = new double[width * height];
         for(int i = 0; i < width; i++) {
             for (int j = 0; j < height; j++) {
                 this.passabilityArray[i + j * width] = (i * j + i + j) / (i * j + 1);
